refactor: createLottoBox의 로직을 구획별로 분리
