+++ src/step2-index.js
@@ -2,10 +2,11 @@ import createGameBox from "./view/web/layers/gameBox/createGameBox.js";
 import createHeader from "./view/web/layers/header/header.js";
 import readLottoPriceInput from "./view/web/modules/readLottoPriceInput.js";
 
-const app = async () => {
-  document.addEventListener("DOMContentLoaded", async () => {
+const app = () => {
+  document.addEventListener("DOMContentLoaded", () => {
     createHeader();
     createGameBox();
+
     readLottoPriceInput();
   });
 };
