+++ src/view/web/layers/lottoBox/createLottoBox.js
@@ -1,38 +1,15 @@
-import { PROMPT_MESSAGE } from "../../../../constants/message.js";
+import createLottoNumbersList from "./createLottoNumbersList.js";
+import createPurchaseMessage from "./createPurchaseMessage.js";
 import "./lottoBox.css";
 
 const createLottoBox = (lottoNumbers) => {
-  const gameBox = document.getElementById("gameBox-container");
   const lottoBox = document.createElement("div");
   lottoBox.id = "lotto-container";
-  gameBox.appendChild(lottoBox);
 
-  const purchaseMessageSpan = document.createElement("span");
-  const purchaseQuantityMessage = `${lottoNumbers.length}${PROMPT_MESSAGE.PURCHASE_QUANTITY}`;
-  purchaseMessageSpan.id = "purchase-quantity-message";
-  purchaseMessageSpan.append(purchaseQuantityMessage);
-  lottoBox.appendChild(purchaseMessageSpan);
+  document.getElementById("gameBox-container").appendChild(lottoBox);
 
-  const ul = document.createElement("ul");
-  ul.className = "lotto-list";
-
-  lottoNumbers.forEach((numbers, index) => {
-    const li = document.createElement("li");
-    li.className = "lotto-numbers";
-    li.id = `lotto-numbers-${index}`;
-
-    const span = document.createElement("span");
-    span.className = "lotto-icon";
-    span.id = `lotto-icon-${index}`;
-    span.append("🎟️");
-
-    const textNode = document.createTextNode(numbers.join(", "));
-
-    li.appendChild(span);
-    li.appendChild(textNode);
-    ul.appendChild(li);
-    lottoBox.appendChild(ul);
-  });
+  createPurchaseMessage(lottoNumbers);
+  createLottoNumbersList(lottoNumbers);
 };
 
 export default createLottoBox;
